Moves GenerateResponse and GenerateResponseChunk to their own files (#1136)
